requirements.txt
triton==2.3.1
timm==1.0.12
mmengine==0.10.5
mmsegmentation==1.2.2
fairscale==0.4.13
yacs==0.1.8
albumentations==1.4.21
albucore==0.0.20
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
Pillow
scipy==1.14.1
numpy==1.26.4
einops==0.8.0
ttach==0.0.3
pandas==2.2.3
matplotlib==3.9.3
seaborn==0.13.2
huggingface-hub==0.26.3
safetensors==0.4.5
PyYAML
tqdm==4.67.1
prettytable==3.12.0
rich==13.9.4
regex==2024.11.6
scipy==1.14.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2